Refactor xlsx_gen
